Rename example and helper directories
