+++ bopo_cli.py
