+++ examples/__init__.py
@@ -1,0 +1,1 @@
+"""Reference examples demonstrating Kermut CLI usage."""
