+++ examples/conditional_probabilities.sh
