+++ examples/conditional_probabilities_all.sh
