+++ examples/conditional_probabilities_single.sh
