+++ helper_scripts/model_adapter.py
